split trait and implementation
