+++ services/src/projects/mod.rs
@@ -1,2 +1,3 @@
+pub mod hashmap_projectdb;
 pub mod project;
 pub mod projectdb;
