+++ services/src/users/mod.rs
@@ -1,3 +1,4 @@
+pub mod hashmap_userdb;
 pub mod session;
 pub mod user;
 pub mod userdb;
